backend/requirements.txt
fastapi
uvicorn
requests
httpx
beautifulsoup4
python-nmap
python-whois
reportlab
weasyprint
matplotlib
plotly
asyncpg
redis
python-dotenv
aiofiles
sqlmodel
psutil
typer
jinja2
